chore: decouple thorswap types from v1 swapper
